requirements.txt
requests==2.28.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
urllib3==1.26.13
colorama==0.4.6
bs4==0.0.1
beautifulsoup4==4.11.1
soupsieve==2.3.2.post1
dnspython==2.2.1

